Remove snap install instructions
Snap makes it very hard to install specific versions of the compiler,
and silently upgrades users to the newest version. While this is
probably a good choice for a lot of desktop applications, this is an
awful property for compilers. Furthermore, we've since:

 * Made a starter project that makes "installing" Clash obsolete
 * Have discovered that we can easily run Clash in non-project mode
   with Stack

I believe we should therefore not advertise Snap installs anymore.