Push users towards starter projects
They are really the best supported flow, and we should make it clear
other methods can have surprises.

The `stack exec` line uses a working Stackage nightly that has a version
of Clash we actually support.

Our Clash starter projects enable `-Wall -Wcompat`. I took the `-Wall`
and added it to the `stack exec` line as well, as it is the most
important one. Adding `-Wcompat` would make the invocation more verbose
without a significant increase in functionality.